Add BulkDeleteWithConfirm
